requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
python-dotenv==1.0.0
Pillow==10.1.0
google-generativeai==0.3.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
